# pkg/api/proto/v1/apirpc.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Quantos Karo Api Service",
    "version": "1.0"
  },
  "tags": [
    {
      "name": "ApiTestService"
    },
    {
      "name": "ApiService"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/getChainInfo": {
      "get": {
        "operationId": "ApiService_GetChainInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ChainInfoResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "type": "string",
              "format": "string"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ApiService"
        ]
      }
    },
    "/getNodeInfo": {
      "get": {
        "operationId": "ApiService_GetNodeInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1NodeInfoResponse"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "type": "string",
              "format": "string"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ApiService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1ApiResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "time": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1ChainInfoResponse": {
      "type": "object",
      "properties": {
        "netName": {
          "type": "string",
          "title": "the name of network, such mainnet or testnet"
        },
        "protocolVersion": {
          "type": "string",
          "title": "the iost protocol version"
        },
        "chainId": {
          "type": "integer",
          "format": "int64",
          "title": "chain id"
        },
        "headBlock": {
          "type": "string",
          "format": "int64",
          "title": "head block height"
        },
        "headBlockHash": {
          "type": "string",
          "title": "head block hash"
        },
        "libBlock": {
          "type": "string",
          "format": "int64",
          "title": "last irreversible block number"
        },
        "libBlockHash": {
          "type": "string",
          "title": "last irreversible block hash"
        },
        "witnessList": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "the current witness list"
        },
        "libWitnessList": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "the active witness list of lat irreversible block"
        },
        "pendingWitnessList": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "the pending witness list"
        },
        "headBlockTime": {
          "type": "string",
          "format": "int64",
          "title": "the head block time"
        },
        "libBlockTime": {
          "type": "string",
          "format": "int64",
          "title": "the last irreversible block time"
        }
      },
      "description": "The message defines chain information response."
    },
    "v1NetworkInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "peerCount": {
          "type": "integer",
          "format": "int32"
        },
        "networkName": {
          "type": "string"
        }
      }
    },
    "v1NodeInfoResponse": {
      "type": "object",
      "properties": {
        "buildTime": {
          "type": "string",
          "title": "build time"
        },
        "gitHash": {
          "type": "string",
          "title": "git hash"
        },
        "mode": {
          "type": "string",
          "title": "node mode"
        },
        "network": {
          "$ref": "#/definitions/v1NetworkInfo",
          "title": "network connection information"
        },
        "codeVersion": {
          "type": "string",
          "title": "code version string"
        },
        "serverTime": {
          "type": "string",
          "format": "int64",
          "title": "server time"
        },
        "time": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}